+++ CHANGELOG.md
@@ -6,6 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+
+## [2024.10.16]
+
+### Added
+
+- Designed interfaces for ConcertSetlist component
+- Created a kernel and extended interface
+
+### Changed
+- Included moveSong, findSong, and getSongAt
+
+
+## [2024.10.02]
+
+### Added
+
+- Designed a proof of concept for ConcertSetlist component
+- Showed all of the basic methods other than moveSong, findSong, and getSongAt
+
 ## [2024.09.16]
 
 ### Added